Move components related to registration into the register folder and update paths that reference the files
