+++ frontend/src/components/register/Register.css
@@ -1,5 +1,5 @@
 html {
-  background-image: url(../../src/images/field_1.jpg);
+  background-image: url(../../../src/images/field_1.jpg);
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
